chore(applicationServices): moved applicationServices to application.services package, docs included
