+++ src/main/java/com/example/roommate/application/services/BookingApplicationService.java
@@ -1,4 +1,4 @@
-package com.example.roommate.applicationServices;
+package com.example.roommate.application.services;
 
 import com.example.roommate.annotations.ApplicationService;
 import com.example.roommate.application.data.BookingData;
