+++ src/test/java/com/example/roommate/tests/services/LoginApplicationServiceTest.java
@@ -3,7 +3,7 @@ package com.example.roommate.tests.services;
 import com.example.roommate.annotations.TestClass;
 import com.example.roommate.factories.ValuesFactory;
 import com.example.roommate.values.forms.LoginForm;
-import com.example.roommate.applicationServices.LoginApplicationService;
+import com.example.roommate.application.services.LoginApplicationService;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
